+++ R/hnsw.R
@@ -86,7 +86,6 @@
 #' Efficient and robust approximate nearest neighbor search using Hierarchical
 #' Navigable Small World graphs.
 #' *arXiv preprint* *arXiv:1603.09320*.
-#' @export
 hnsw_knn <- function(X,
                      k = 10,
                      distance = "euclidean",
@@ -180,7 +179,6 @@ hnsw_knn <- function(X,
 #' irism <- as.matrix(iris[, -5])
 #' ann <- hnsw_build(irism)
 #' iris_nn <- hnsw_search(irism, ann, k = 5)
-#' @export
 hnsw_build <- function(X,
                        distance = "euclidean",
                        M = 16,
@@ -299,7 +297,6 @@ hnsw_build <- function(X,
 #' irism <- as.matrix(iris[, -5])
 #' ann <- hnsw_build(irism)
 #' iris_nn <- hnsw_search(irism, ann, k = 5)
-#' @export
 hnsw_search <-
   function(X,
            ann,